# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Social Badge API
  description: |
    Backend API for the Social Badge platform — a digital badge builder that enables
    event organisers to create customisable badge templates and participants to
    personalise and share badges on social media.
  version: 1.0.0
  contact:
    name: Social Badge Team
servers:
  - url: http://127.0.0.1:8000/api/v1
    description: Local development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /auth/login or /auth/signup
  schemas:
    SignupRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: organiser@example.com
        password:
          type: string
          minLength: 8
          example: S3cureP@ssword
...
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
          example: Resource not found
    ValidationErrorResponse:
      type: object
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              loc:
                type: array
                items:
                  type: string
              msg:
                type: string
              type:
                type: string
paths:
...
  /auth/signup:
    post:
      tags: [Auth]
      summary: Register a new organiser account
      description: |
        Creates a new organiser account with email and password.
 
        - Password is hashed with a minimum of 10 salt rounds; plaintext is
          never stored.
        - Sets `email_verified = false` on the new account and sends a
          one-time verification email with a 30-minute TTL token.
        - Returns `409` if the email already exists — the error message must
          not reveal that the address is registered (no email enumeration).
        - All input is validated server-side regardless of client-side validation.
        - Rate-limited to **10 requests per IP per minute**.
        - All error responses use `{ "status": "error", "message": "..." }`.
      operationId: signup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
      responses:
        "201":
          description: >
            Account created. Verification email sent. No token returned in body.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: Verification email sent.
                  data:
                    type: object
                    example: {}
        "400":
          description: Validation error — missing field, invalid email format, or weak password
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: "Password must be at least 8 characters and include an uppercase letter, a number, and a special character."
        "409":
          description: >
            Email already exists. Message deliberately vague to prevent
            email enumeration.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: "Unable to create account. Please try a different email or log in."
        "429":
          description: Rate limit exceeded — 10 requests per IP per minute
          headers:
            Retry-After:
              schema:
                type: integer
                example: 60
            X-RateLimit-Limit:
              schema:
                type: integer
                example: 10
            X-RateLimit-Remaining:
              schema:
                type: integer
                example: 0
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: "Too many requests. Please try again later."
        "500":
          description: Unexpected internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: "An unexpected error occurred. Please try again."
...
  /auth/forgot-password:
    post:
      tags: [Auth]
      summary: Request a password reset link
      description: |
        Accepts an email address and sends a password reset link containing a
        one-time token to that address.
        - Always returns `200 OK` regardless of whether the email exists, to prevent email enumeration attacks.
        - If the email is found: generates a cryptographically random one-time token, stores a hashed copy with a 30-minute expiry, and sends a reset email with the raw token as a query parameter (e.g. `https://socialbadge.io/reset-password?token=<raw_token>`).
        - If the email is not found: no email is sent, but the same `200` response is returned.
        - Rate-limited to **10 requests per IP per minute**; exceeding this returns `429 Too Many Requests`.
      operationId: forgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
                  example: organiser@example.com
      responses:
        "200":
          description: >
            Request received. If the email is registered, a reset link has been
            sent. No indication is given as to whether the email exists.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: If that email is registered, a reset link has been sent.
        "400":
          description: Invalid request body (e.g. malformed email format)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        "429":
          description: Rate limit exceeded — maximum 10 requests per IP per minute
          headers:
            Retry-After:
              description: Number of seconds to wait before retrying
              schema:
                type: integer
                example: 60
            X-RateLimit-Limit:
              description: Maximum requests allowed per minute
              schema:
                type: integer
                example: 10
            X-RateLimit-Remaining:
              description: Remaining requests in the current window
              schema:
                type: integer
                example: 0
          content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorResponse'
               

    /auth/reset-password:
      post:
        tags: [Auth]
        summary: Reset password using a one-time token
        description: 
        - Validates the reset token and updates the user's password.
        - Accepts a reset token, new password, and password confirmation.
          Password match validation is enforced on both frontend and backend.
        - Token must exist, not be expired (1-hour TTL), and not have been
          previously used.
        - New password must meet the same strength requirements as signup.
        - On success:
          - Hashes and saves the new password to `users.password_hash`
          - Marks the reset token as `used = true`
          - Revokes **all** refresh tokens for the user, forcing re-login on
            all devices
        - Returns `400` with `"Token is invalid or expired"` if validation fails.
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, new_password, confirm_password]
              properties:
                token:
                  type: string
                  description: Raw one-time reset token received via email link
                  example: a3f9c2e1b7d04a8e9f2c3b1d6e5a7f04
                new_password:
                  type: string
                  minLength: 8
                  description: New password (must meet signup strength requirements)
                  example: NewS3cure!Pass
                confirm_password:
                  type: string
                  description: Must match new_password exactly
                  example: NewS3cure!Pass
      responses:
        "200":
          description: Password reset successful — user should proceed to login
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: Password reset successful. Please log in with your new password.
        "400":
          description: |
            One of the following:
            - Token is invalid, expired, or already used
            - Passwords do not match
            - New password does not meet strength requirements
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: Token is invalid or expired
  /auth/verify-email:
    post:
      tags: [Auth]
      summary: Verify a new organiser's email address
      description: |
        Validates a one-time email verification token and activates the user account.
 
        - Token is passed as a query parameter (e.g. from the link in the
          verification email).
        - Token must exist, be linked to a pending user account, and not be
          expired (30-minute TTL).
        - On success: sets `email_verified = true` and immediately invalidates
          the token.
        - All error responses use the format:
          `{ "status": "error", "message": "<error message>" }`
      operationId: verifyEmail
      parameters:
        - name: token
          in: query
          required: true
          description: One-time verification token received via email
          schema:
            type: string
            example: b2e7f1c4a09d3e6b8f2a1c5d9e4b7f03
      responses:
        "200":
          description: Email verified successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: Email verified
                  data:
                    type: object
                    properties:
                      next:
                        type: string
                        example: onboarding
        "400":
          description: Account is already verified
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: User already verified
        "401":
          description: Token has expired
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: "Token has expired. Please request a new verification email."
...